Allow map key and value types to be specified.
Map type hint objects now include `keyType` and `valueType` properties.
Because JavaScript object properties are automatically coerced to strings,
`keyType` is unlikely to be useful beyond `"text"`.

Optionally, type hints passed to `encoder.encode` may now be specified
as strings, constants (from `types.dataTypes`), and objects. If the
hint is an object, it will be assumed to contain explicit type and
subtype information.